Refactor solver and add CLI with documentation
